refactor: Modularize frame display system into focused components
Refactored the monolithic frame_disp.h (~3,500
  lines) into a clean
  modular architecture with 15 focused files for
  better maintainability.

  Changes:
  - Split frame display into 6 core modules (IF,
  EF, AS, AT, AF, common)
  - Extracted 16+ effect types into 9 dedicated
  effect handlers
  - Organized effect modules by category in
  frame_disp/effects/
  - Consolidated frame_disp.h as single entry point
  - Removed backup files and intermediate
  refactoring artifacts

  Structure:
  - src/frame_disp.h - Main header (30 lines)
  - src/frame_disp/*.h - Core display modules (5
  files)
  - src/frame_disp/effects/*.h - Effect type
  handlers (9 files)

  Total: 3,718 lines across 15 well-organized
  modules
