+++ src/frame_disp/frame_disp_af.h
@@ -1,0 +1,254 @@
+#ifndef FRAME_DISP_AF_H_GUARD
+#define FRAME_DISP_AF_H_GUARD
+
+#include "frame_disp_common.h"
+#include "../cg.h"
+#include <cstring> // for memcpy
+
+// ============================================================================
+// Animation Frame (AF) Display  
+// ============================================================================
+// Displays and edits Frame_AF data (sprite, animation, transforms)
+//
+// Frame_AF controls:
+// - Sprite selection and .pat usage
+// - Animation control (duration, loops, jumps, landing frames)
+// - Z-priority
+// - Visual transforms (position, rotation, scale, color, blend mode)
+// - Interpolation
+// ============================================================================
+
+inline void AfDisplay(Frame_AF *af, int &selectedLayer, FrameData *frameData = nullptr, int patternIndex = -1, std::function<void()> onModified = nullptr)
+{
+	// Helper lambda to mark both frameData and character as modified
+	auto markModified = [&]() {
+		if (frameData && patternIndex >= 0) {
+			frameData->mark_modified(patternIndex);
+		}
+		if (onModified) {
+			onModified();
+		}
+	};
+
+
+	constexpr float width = 50.f;
+
+	// Copy/paste buttons for animation section
+	static Frame_AF copiedAnimation = {};
+	if(im::SmallButton("Copy animation")) {
+		copiedAnimation.spriteId = af->spriteId;
+		copiedAnimation.usePat = af->usePat;
+		copiedAnimation.duration = af->duration;
+		copiedAnimation.aniType = af->aniType;
+		copiedAnimation.aniFlag = af->aniFlag;
+		copiedAnimation.jump = af->jump;
+		copiedAnimation.landJump = af->landJump;
+		copiedAnimation.priority = af->priority;
+		copiedAnimation.loopCount = af->loopCount;
+		copiedAnimation.loopEnd = af->loopEnd;
+	}
+	if(im::IsItemHovered()) Tooltip("Copy sprite, duration, jumps, priority, and loops");
+	im::SameLine();
+	if(im::SmallButton("Paste animation")) {
+		af->spriteId = copiedAnimation.spriteId;
+		af->usePat = copiedAnimation.usePat;
+		af->duration = copiedAnimation.duration;
+		af->aniType = copiedAnimation.aniType;
+		af->aniFlag = copiedAnimation.aniFlag;
+		af->jump = copiedAnimation.jump;
+		af->landJump = copiedAnimation.landJump;
+		af->priority = copiedAnimation.priority;
+		af->loopCount = copiedAnimation.loopCount;
+		af->loopEnd = copiedAnimation.loopEnd;
+		markModified();
+	}
+	if(im::IsItemHovered()) Tooltip("Paste sprite, duration, jumps, priority, and loops");
+
+	// MBAACC uses flat structure (no layers)
+	im::SetNextItemWidth(width*3);
+	if(im::InputInt("Sprite", &af->spriteId)) {
+		markModified();
+	}
+	im::SameLine(0, 20.f);
+	if(im::Checkbox("Use .pat", &af->usePat)) {
+		markModified();
+	}
+
+	im::Separator();
+
+	unsigned int flagIndex = -1;
+	if(BitField("Animation flags", &af->aniFlag, &flagIndex, 4)) {
+		markModified();
+	}
+	switch (flagIndex)
+	{
+		case 0: Tooltip("Landing frame: Land to pattern"); break;
+		case 1: Tooltip("Loop: Decrement loops and go to end if 0"); break;
+		case 2: Tooltip("Go to: Use relative offset"); break;
+		case 3: Tooltip("End of loop: Use relative offset"); break;
+	}
+
+	if(im::Combo("Animation", &af->aniType, animationList, IM_ARRAYSIZE(animationList))) {
+		markModified();
+	}
+
+	im::SetNextItemWidth(width);
+	if(im::InputInt("Go to", &af->jump, 0, 0)) {
+		markModified();
+	}
+	im::SameLine(0.f, 20); im::SetNextItemWidth(width);
+	if(im::InputInt("Landing frame", &af->landJump, 0, 0)) {
+		markModified();
+	}
+
+	// Landing frame tools
+	if(frameData && patternIndex >= 0) {
+		static int landingFrameToolValue = 0;
+		static int landingFrameRange[2] = {0, 0};
+
+		im::SetNextItemWidth(width);
+		im::InputInt("##LandingValue", &landingFrameToolValue, 0, 0);
+		if(im::IsItemHovered()) Tooltip("Landing frame value to set");
+
+		im::SameLine();
+		if(im::SmallButton("Set All")) {
+			auto seq = frameData->get_sequence(patternIndex);
+			if(seq) {
+				for(int i = 0; i < seq->frames.size(); i++) {
+					seq->frames[i].AF.landJump = landingFrameToolValue;
+				}
+				frameData->mark_modified(patternIndex);
+				markModified();
+			}
+		}
+		if(im::IsItemHovered()) Tooltip("Apply landing frame to all frames in pattern");
+
+		im::SameLine();
+		if(im::SmallButton("Set Range")) {
+			im::OpenPopup("LandingFrameRange");
+		}
+		if(im::IsItemHovered()) Tooltip("Apply landing frame to a range of frames");
+
+		// Range popup
+		if(im::BeginPopup("LandingFrameRange")) {
+			auto seq = frameData->get_sequence(patternIndex);
+			if(seq) {
+				const int maxFrame = seq->frames.size() - 1;
+
+				// Clamp range values
+				if(landingFrameRange[0] < 0) landingFrameRange[0] = 0;
+				if(landingFrameRange[1] < 0) landingFrameRange[1] = 0;
+				if(landingFrameRange[0] > maxFrame) landingFrameRange[0] = maxFrame;
+				if(landingFrameRange[1] > maxFrame) landingFrameRange[1] = maxFrame;
+
+				im::Text("Set landing frame for range");
+				im::Separator();
+				im::InputInt2("Frame range", landingFrameRange);
+				im::InputInt("Landing frame value", &landingFrameToolValue);
+
+				if(im::Button("Apply", ImVec2(120, 0))) {
+					for(int i = landingFrameRange[0]; i <= landingFrameRange[1] && i >= 0 && i < seq->frames.size(); i++) {
+						seq->frames[i].AF.landJump = landingFrameToolValue;
+					}
+					frameData->mark_modified(patternIndex);
+					markModified();
+					im::CloseCurrentPopup();
+				}
+				im::SameLine();
+				if(im::Button("Cancel", ImVec2(120, 0))) {
+					im::CloseCurrentPopup();
+				}
+			}
+			im::EndPopup();
+		}
+	}
+
+	im::SetNextItemWidth(width);
+	if(im::InputInt("Z-Priority", &af->priority, 0, 0)) {
+		markModified();
+	}
+	im::SetNextItemWidth(width);
+	if(im::InputInt("Loop N times", &af->loopCount, 0, 0)) {
+		markModified();
+	}
+	im::SameLine(0,20); im::SetNextItemWidth(width);
+	if(im::InputInt("End of loop", &af->loopEnd, 0, 0)) {
+		markModified();
+	}
+	if(im::InputInt("Duration", &af->duration, 1, 0)) {
+		markModified();
+	}
+
+	im::Separator();
+
+	// Copy/paste buttons for transform section
+	static Frame_AF copiedTransform = {};
+	if(im::SmallButton("Copy transforms")) {
+		copiedTransform.offset_x = af->offset_x;
+		copiedTransform.offset_y = af->offset_y;
+		copiedTransform.blend_mode = af->blend_mode;
+		memcpy(copiedTransform.rgba, af->rgba, sizeof(af->rgba));
+		memcpy(copiedTransform.rotation, af->rotation, sizeof(af->rotation));
+		memcpy(copiedTransform.scale, af->scale, sizeof(af->scale));
+		copiedTransform.AFRT = af->AFRT;
+		copiedTransform.interpolationType = af->interpolationType;
+	}
+	if(im::IsItemHovered()) Tooltip("Copy offset, rotation, scale, color, blend mode, and interpolation");
+	im::SameLine();
+	if(im::SmallButton("Paste transforms")) {
+		af->offset_x = copiedTransform.offset_x;
+		af->offset_y = copiedTransform.offset_y;
+		af->blend_mode = copiedTransform.blend_mode;
+		memcpy(af->rgba, copiedTransform.rgba, sizeof(af->rgba));
+		memcpy(af->rotation, copiedTransform.rotation, sizeof(af->rotation));
+		memcpy(af->scale, copiedTransform.scale, sizeof(af->scale));
+		af->AFRT = copiedTransform.AFRT;
+		af->interpolationType = copiedTransform.interpolationType;
+		markModified();
+	}
+	if(im::IsItemHovered()) Tooltip("Paste offset, rotation, scale, color, blend mode, and interpolation");
+
+	if(im::Combo("Interpolation", &af->interpolationType, interpolationList, IM_ARRAYSIZE(interpolationList))) {
+		markModified();
+	}
+
+	im::SetNextItemWidth(width);
+	im::DragInt("X", &af->offset_x);
+	if(im::IsItemDeactivatedAfterEdit()) {
+		markModified();
+	}
+	im::SameLine();
+	im::SetNextItemWidth(width);
+	im::DragInt("Y", &af->offset_y);
+	if(im::IsItemDeactivatedAfterEdit()) {
+		markModified();
+	}
+
+	int mode = af->blend_mode-1;
+	if(mode < 1)
+		mode = 0;
+	if (im::Combo("Blend Mode", &mode, "Normal\0Additive\0Subtractive\0"))
+	{
+		af->blend_mode=mode+1;
+		markModified();
+	}
+	if(im::ColorEdit4("Color", af->rgba)) {
+		markModified();
+	}
+
+	im::DragFloat3("Rot XYZ", af->rotation, 0.005);
+	if(im::IsItemDeactivatedAfterEdit()) {
+		markModified();
+	}
+	im::DragFloat2("Scale", af->scale, 0.1);
+	if(im::IsItemDeactivatedAfterEdit()) {
+		markModified();
+	}
+	if(im::Checkbox("Rotation keeps scale set by EF", &af->AFRT)) {
+		markModified();
+	}
+
+}
+
+#endif /* FRAME_DISP_AF_H_GUARD */
+
